P5LR5/web_gateway/requirements.txt
flask~=2.3.0
grpcio~=1.60.0
grpcio-tools~=1.60.0  
Jinja2~=3.1.0